feat(server): refactor node dtos
