+++ server/src/nodes/dtos/commons/index.ts
@@ -1,0 +1,1 @@
+export { NodeInfoDto } from './node-info.dto';
