+++ server/src/nodes/dtos/commons/node-info.dto.ts
