--- server/src/nodes/dtos/create-node.dto.ts
@@ -1,4 +1,0 @@
-import { OmitType } from '@nestjs/swagger';
-import { NodeDto } from './node.dto';
-
-export class CreateNodeDto extends OmitType(NodeDto, ['nodeId']) {}
