+++ server/src/nodes/dtos/requests/create-node-request.dto.ts
@@ -1,0 +1,4 @@
+import { OmitType } from '@nestjs/swagger';
+import { NodeDto } from '../node.dto';
+
+export class CreateNodeRequestDto extends OmitType(NodeDto, ['id'] as const) {}
