+++ server/src/nodes/dtos/requests/index.ts
@@ -1,0 +1,2 @@
+export { CreateNodeRequestDto } from './create-node-request.dto';
+export { UpdateNodeRequestDto } from './update-node-request.dto';
