+++ server/src/nodes/dtos/requests/update-node-request.dto.ts
@@ -1,0 +1,6 @@
+import { OmitType, PartialType } from '@nestjs/swagger';
+import { NodeDto } from '../node.dto';
+
+class NodeWithoutId extends OmitType(NodeDto, ['id'] as const) {}
+
+export class UpdateNodeRequestDto extends PartialType(NodeWithoutId) {}
