+++ server/src/nodes/dtos/responses/index.ts
@@ -1,0 +1,1 @@
+export { NodeResponseDto } from './node-response.dto';
