+++ server/src/nodes/dtos/responses/node-response.dto.ts
@@ -1,0 +1,3 @@
+import { NodeDto } from '../node.dto';
+
+export class NodeResponseDto extends NodeDto {}
